+++ internal/cloud/aws_test.go
@@ -127,7 +127,7 @@ func TestAWSCloud_Send(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockUploadArchive: func(*glacier.UploadArchiveInput) (*glacier.ArchiveCreationOutput, error) {
 						return &glacier.ArchiveCreationOutput{
 							ArchiveId: aws.String("AWSID123"),
@@ -166,7 +166,7 @@ func TestAWSCloud_Send(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockUploadArchive: func(*glacier.UploadArchiveInput) (*glacier.ArchiveCreationOutput, error) {
 						return nil, errors.New("connection error")
 					},
@@ -196,7 +196,7 @@ func TestAWSCloud_Send(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockUploadArchive: func(*glacier.UploadArchiveInput) (*glacier.ArchiveCreationOutput, error) {
 						return &glacier.ArchiveCreationOutput{
 							ArchiveId: aws.String("AWSID123"),
@@ -230,7 +230,7 @@ func TestAWSCloud_Send(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateMultipartUpload: func(i *glacier.InitiateMultipartUploadInput) (*glacier.InitiateMultipartUploadOutput, error) {
 						partSize, err := strconv.ParseInt(*i.PartSize, 10, 64)
 						if err != nil {
@@ -291,7 +291,7 @@ func TestAWSCloud_Send(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateMultipartUpload: func(*glacier.InitiateMultipartUploadInput) (*glacier.InitiateMultipartUploadOutput, error) {
 						return nil, errors.New("aws is out")
 					},
@@ -321,7 +321,7 @@ func TestAWSCloud_Send(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockAbortMultipartUpload: func(*glacier.AbortMultipartUploadInput) (*glacier.AbortMultipartUploadOutput, error) {
 						return nil, nil
 					},
@@ -370,7 +370,7 @@ func TestAWSCloud_Send(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockAbortMultipartUpload: func(*glacier.AbortMultipartUploadInput) (*glacier.AbortMultipartUploadOutput, error) {
 						return nil, nil
 					},
@@ -410,7 +410,7 @@ func TestAWSCloud_Send(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockAbortMultipartUpload: func(*glacier.AbortMultipartUploadInput) (*glacier.AbortMultipartUploadOutput, error) {
 						return nil, nil
 					},
@@ -454,7 +454,7 @@ func TestAWSCloud_Send(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockDeleteArchive: func(d *glacier.DeleteArchiveInput) (*glacier.DeleteArchiveOutput, error) {
 						if *d.ArchiveId != "UPLOAD123" {
 							return nil, fmt.Errorf("unexpected id %s", *d.ArchiveId)
@@ -506,7 +506,7 @@ func TestAWSCloud_Send(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockDeleteArchive: func(*glacier.DeleteArchiveInput) (*glacier.DeleteArchiveOutput, error) {
 						return nil, errors.New("connection error")
 					},
@@ -570,7 +570,7 @@ func TestAWSCloud_List(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -642,7 +642,7 @@ func TestAWSCloud_List(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return nil, errors.New("a crazy error")
 					},
@@ -655,7 +655,7 @@ func TestAWSCloud_List(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -673,7 +673,7 @@ func TestAWSCloud_List(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -700,7 +700,7 @@ func TestAWSCloud_List(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -726,7 +726,7 @@ func TestAWSCloud_List(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -789,7 +789,7 @@ func TestAWSCloud_List(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -818,7 +818,7 @@ func TestAWSCloud_List(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -878,7 +878,7 @@ func TestAWSCloud_Get(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -910,7 +910,7 @@ func TestAWSCloud_Get(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return nil, errors.New("a crazy error")
 					},
@@ -924,7 +924,7 @@ func TestAWSCloud_Get(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -943,7 +943,7 @@ func TestAWSCloud_Get(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -971,7 +971,7 @@ func TestAWSCloud_Get(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -998,7 +998,7 @@ func TestAWSCloud_Get(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -1034,7 +1034,7 @@ func TestAWSCloud_Get(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockInitiateJob: func(*glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 						return &glacier.InitiateJobOutput{
 							JobId: aws.String("JOBID123"),
@@ -1086,7 +1086,7 @@ func TestAWSCloud_Remove(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockDeleteArchive: func(*glacier.DeleteArchiveInput) (*glacier.DeleteArchiveOutput, error) {
 						return &glacier.DeleteArchiveOutput{}, nil
 					},
@@ -1099,7 +1099,7 @@ func TestAWSCloud_Remove(t *testing.T) {
 			awsCloud: cloud.AWSCloud{
 				AccountID: "account",
 				VaultName: "vault",
-				Glacier: glacierAPIMock{
+				Glacier: mockGlacierAPI{
 					mockDeleteArchive: func(*glacier.DeleteArchiveInput) (*glacier.DeleteArchiveOutput, error) {
 						return nil, errors.New("no backup here")
 					},
@@ -1119,7 +1119,7 @@ func TestAWSCloud_Remove(t *testing.T) {
 	}
 }
 
-type glacierAPIMock struct {
+type mockGlacierAPI struct {
 	mockAbortMultipartUploadRequest     func(*glacier.AbortMultipartUploadInput) (*request.Request, *glacier.AbortMultipartUploadOutput)
 	mockAbortMultipartUpload            func(*glacier.AbortMultipartUploadInput) (*glacier.AbortMultipartUploadOutput, error)
 	mockAbortVaultLockRequest           func(*glacier.AbortVaultLockInput) (*request.Request, *glacier.AbortVaultLockOutput)
@@ -1190,275 +1190,275 @@ type glacierAPIMock struct {
 	mockWaitUntilVaultNotExists         func(*glacier.DescribeVaultInput) error
 }
 
-func (g glacierAPIMock) AbortMultipartUploadRequest(a *glacier.AbortMultipartUploadInput) (*request.Request, *glacier.AbortMultipartUploadOutput) {
+func (g mockGlacierAPI) AbortMultipartUploadRequest(a *glacier.AbortMultipartUploadInput) (*request.Request, *glacier.AbortMultipartUploadOutput) {
 	return g.mockAbortMultipartUploadRequest(a)
 }
 
-func (g glacierAPIMock) AbortMultipartUpload(a *glacier.AbortMultipartUploadInput) (*glacier.AbortMultipartUploadOutput, error) {
+func (g mockGlacierAPI) AbortMultipartUpload(a *glacier.AbortMultipartUploadInput) (*glacier.AbortMultipartUploadOutput, error) {
 	return g.mockAbortMultipartUpload(a)
 }
 
-func (g glacierAPIMock) AbortVaultLockRequest(a *glacier.AbortVaultLockInput) (*request.Request, *glacier.AbortVaultLockOutput) {
+func (g mockGlacierAPI) AbortVaultLockRequest(a *glacier.AbortVaultLockInput) (*request.Request, *glacier.AbortVaultLockOutput) {
 	return g.mockAbortVaultLockRequest(a)
 }
 
-func (g glacierAPIMock) AbortVaultLock(a *glacier.AbortVaultLockInput) (*glacier.AbortVaultLockOutput, error) {
+func (g mockGlacierAPI) AbortVaultLock(a *glacier.AbortVaultLockInput) (*glacier.AbortVaultLockOutput, error) {
 	return g.mockAbortVaultLock(a)
 }
 
-func (g glacierAPIMock) AddTagsToVaultRequest(a *glacier.AddTagsToVaultInput) (*request.Request, *glacier.AddTagsToVaultOutput) {
+func (g mockGlacierAPI) AddTagsToVaultRequest(a *glacier.AddTagsToVaultInput) (*request.Request, *glacier.AddTagsToVaultOutput) {
 	return g.mockAddTagsToVaultRequest(a)
 }
 
-func (g glacierAPIMock) AddTagsToVault(a *glacier.AddTagsToVaultInput) (*glacier.AddTagsToVaultOutput, error) {
+func (g mockGlacierAPI) AddTagsToVault(a *glacier.AddTagsToVaultInput) (*glacier.AddTagsToVaultOutput, error) {
 	return g.mockAddTagsToVault(a)
 }
 
-func (g glacierAPIMock) CompleteMultipartUploadRequest(c *glacier.CompleteMultipartUploadInput) (*request.Request, *glacier.ArchiveCreationOutput) {
+func (g mockGlacierAPI) CompleteMultipartUploadRequest(c *glacier.CompleteMultipartUploadInput) (*request.Request, *glacier.ArchiveCreationOutput) {
 	return g.mockCompleteMultipartUploadRequest(c)
 }
 
-func (g glacierAPIMock) CompleteMultipartUpload(c *glacier.CompleteMultipartUploadInput) (*glacier.ArchiveCreationOutput, error) {
+func (g mockGlacierAPI) CompleteMultipartUpload(c *glacier.CompleteMultipartUploadInput) (*glacier.ArchiveCreationOutput, error) {
 	return g.mockCompleteMultipartUpload(c)
 }
 
-func (g glacierAPIMock) CompleteVaultLockRequest(c *glacier.CompleteVaultLockInput) (*request.Request, *glacier.CompleteVaultLockOutput) {
+func (g mockGlacierAPI) CompleteVaultLockRequest(c *glacier.CompleteVaultLockInput) (*request.Request, *glacier.CompleteVaultLockOutput) {
 	return g.mockCompleteVaultLockRequest(c)
 }
 
-func (g glacierAPIMock) CompleteVaultLock(c *glacier.CompleteVaultLockInput) (*glacier.CompleteVaultLockOutput, error) {
+func (g mockGlacierAPI) CompleteVaultLock(c *glacier.CompleteVaultLockInput) (*glacier.CompleteVaultLockOutput, error) {
 	return g.mockCompleteVaultLock(c)
 }
 
-func (g glacierAPIMock) CreateVaultRequest(c *glacier.CreateVaultInput) (*request.Request, *glacier.CreateVaultOutput) {
+func (g mockGlacierAPI) CreateVaultRequest(c *glacier.CreateVaultInput) (*request.Request, *glacier.CreateVaultOutput) {
 	return g.mockCreateVaultRequest(c)
 }
 
-func (g glacierAPIMock) CreateVault(c *glacier.CreateVaultInput) (*glacier.CreateVaultOutput, error) {
+func (g mockGlacierAPI) CreateVault(c *glacier.CreateVaultInput) (*glacier.CreateVaultOutput, error) {
 	return g.mockCreateVault(c)
 }
 
-func (g glacierAPIMock) DeleteArchiveRequest(d *glacier.DeleteArchiveInput) (*request.Request, *glacier.DeleteArchiveOutput) {
+func (g mockGlacierAPI) DeleteArchiveRequest(d *glacier.DeleteArchiveInput) (*request.Request, *glacier.DeleteArchiveOutput) {
 	return g.mockDeleteArchiveRequest(d)
 }
 
-func (g glacierAPIMock) DeleteArchive(d *glacier.DeleteArchiveInput) (*glacier.DeleteArchiveOutput, error) {
+func (g mockGlacierAPI) DeleteArchive(d *glacier.DeleteArchiveInput) (*glacier.DeleteArchiveOutput, error) {
 	return g.mockDeleteArchive(d)
 }
 
-func (g glacierAPIMock) DeleteVaultRequest(d *glacier.DeleteVaultInput) (*request.Request, *glacier.DeleteVaultOutput) {
+func (g mockGlacierAPI) DeleteVaultRequest(d *glacier.DeleteVaultInput) (*request.Request, *glacier.DeleteVaultOutput) {
 	return g.mockDeleteVaultRequest(d)
 }
 
-func (g glacierAPIMock) DeleteVault(d *glacier.DeleteVaultInput) (*glacier.DeleteVaultOutput, error) {
+func (g mockGlacierAPI) DeleteVault(d *glacier.DeleteVaultInput) (*glacier.DeleteVaultOutput, error) {
 	return g.mockDeleteVault(d)
 }
 
-func (g glacierAPIMock) DeleteVaultAccessPolicyRequest(d *glacier.DeleteVaultAccessPolicyInput) (*request.Request, *glacier.DeleteVaultAccessPolicyOutput) {
+func (g mockGlacierAPI) DeleteVaultAccessPolicyRequest(d *glacier.DeleteVaultAccessPolicyInput) (*request.Request, *glacier.DeleteVaultAccessPolicyOutput) {
 	return g.mockDeleteVaultAccessPolicyRequest(d)
 }
 
-func (g glacierAPIMock) DeleteVaultAccessPolicy(d *glacier.DeleteVaultAccessPolicyInput) (*glacier.DeleteVaultAccessPolicyOutput, error) {
+func (g mockGlacierAPI) DeleteVaultAccessPolicy(d *glacier.DeleteVaultAccessPolicyInput) (*glacier.DeleteVaultAccessPolicyOutput, error) {
 	return g.mockDeleteVaultAccessPolicy(d)
 }
 
-func (g glacierAPIMock) DeleteVaultNotificationsRequest(d *glacier.DeleteVaultNotificationsInput) (*request.Request, *glacier.DeleteVaultNotificationsOutput) {
+func (g mockGlacierAPI) DeleteVaultNotificationsRequest(d *glacier.DeleteVaultNotificationsInput) (*request.Request, *glacier.DeleteVaultNotificationsOutput) {
 	return g.mockDeleteVaultNotificationsRequest(d)
 }
 
-func (g glacierAPIMock) DeleteVaultNotifications(d *glacier.DeleteVaultNotificationsInput) (*glacier.DeleteVaultNotificationsOutput, error) {
+func (g mockGlacierAPI) DeleteVaultNotifications(d *glacier.DeleteVaultNotificationsInput) (*glacier.DeleteVaultNotificationsOutput, error) {
 	return g.mockDeleteVaultNotifications(d)
 }
 
-func (g glacierAPIMock) DescribeJobRequest(d *glacier.DescribeJobInput) (*request.Request, *glacier.JobDescription) {
+func (g mockGlacierAPI) DescribeJobRequest(d *glacier.DescribeJobInput) (*request.Request, *glacier.JobDescription) {
 	return g.mockDescribeJobRequest(d)
 }
 
-func (g glacierAPIMock) DescribeJob(d *glacier.DescribeJobInput) (*glacier.JobDescription, error) {
+func (g mockGlacierAPI) DescribeJob(d *glacier.DescribeJobInput) (*glacier.JobDescription, error) {
 	return g.mockDescribeJob(d)
 }
 
-func (g glacierAPIMock) DescribeVaultRequest(d *glacier.DescribeVaultInput) (*request.Request, *glacier.DescribeVaultOutput) {
+func (g mockGlacierAPI) DescribeVaultRequest(d *glacier.DescribeVaultInput) (*request.Request, *glacier.DescribeVaultOutput) {
 	return g.mockDescribeVaultRequest(d)
 }
 
-func (g glacierAPIMock) DescribeVault(d *glacier.DescribeVaultInput) (*glacier.DescribeVaultOutput, error) {
+func (g mockGlacierAPI) DescribeVault(d *glacier.DescribeVaultInput) (*glacier.DescribeVaultOutput, error) {
 	return g.mockDescribeVault(d)
 }
 
-func (g glacierAPIMock) GetDataRetrievalPolicyRequest(d *glacier.GetDataRetrievalPolicyInput) (*request.Request, *glacier.GetDataRetrievalPolicyOutput) {
+func (g mockGlacierAPI) GetDataRetrievalPolicyRequest(d *glacier.GetDataRetrievalPolicyInput) (*request.Request, *glacier.GetDataRetrievalPolicyOutput) {
 	return g.mockGetDataRetrievalPolicyRequest(d)
 }
 
-func (g glacierAPIMock) GetDataRetrievalPolicy(d *glacier.GetDataRetrievalPolicyInput) (*glacier.GetDataRetrievalPolicyOutput, error) {
+func (g mockGlacierAPI) GetDataRetrievalPolicy(d *glacier.GetDataRetrievalPolicyInput) (*glacier.GetDataRetrievalPolicyOutput, error) {
 	return g.mockGetDataRetrievalPolicy(d)
 }
 
-func (g glacierAPIMock) GetJobOutputRequest(d *glacier.GetJobOutputInput) (*request.Request, *glacier.GetJobOutputOutput) {
+func (g mockGlacierAPI) GetJobOutputRequest(d *glacier.GetJobOutputInput) (*request.Request, *glacier.GetJobOutputOutput) {
 	return g.mockGetJobOutputRequest(d)
 }
 
-func (g glacierAPIMock) GetJobOutput(d *glacier.GetJobOutputInput) (*glacier.GetJobOutputOutput, error) {
+func (g mockGlacierAPI) GetJobOutput(d *glacier.GetJobOutputInput) (*glacier.GetJobOutputOutput, error) {
 	return g.mockGetJobOutput(d)
 }
 
-func (g glacierAPIMock) GetVaultAccessPolicyRequest(d *glacier.GetVaultAccessPolicyInput) (*request.Request, *glacier.GetVaultAccessPolicyOutput) {
+func (g mockGlacierAPI) GetVaultAccessPolicyRequest(d *glacier.GetVaultAccessPolicyInput) (*request.Request, *glacier.GetVaultAccessPolicyOutput) {
 	return g.mockGetVaultAccessPolicyRequest(d)
 }
 
-func (g glacierAPIMock) GetVaultAccessPolicy(d *glacier.GetVaultAccessPolicyInput) (*glacier.GetVaultAccessPolicyOutput, error) {
+func (g mockGlacierAPI) GetVaultAccessPolicy(d *glacier.GetVaultAccessPolicyInput) (*glacier.GetVaultAccessPolicyOutput, error) {
 	return g.mockGetVaultAccessPolicy(d)
 }
 
-func (g glacierAPIMock) GetVaultLockRequest(d *glacier.GetVaultLockInput) (*request.Request, *glacier.GetVaultLockOutput) {
+func (g mockGlacierAPI) GetVaultLockRequest(d *glacier.GetVaultLockInput) (*request.Request, *glacier.GetVaultLockOutput) {
 	return g.mockGetVaultLockRequest(d)
 }
 
-func (g glacierAPIMock) GetVaultLock(d *glacier.GetVaultLockInput) (*glacier.GetVaultLockOutput, error) {
+func (g mockGlacierAPI) GetVaultLock(d *glacier.GetVaultLockInput) (*glacier.GetVaultLockOutput, error) {
 	return g.mockGetVaultLock(d)
 }
 
-func (g glacierAPIMock) GetVaultNotificationsRequest(d *glacier.GetVaultNotificationsInput) (*request.Request, *glacier.GetVaultNotificationsOutput) {
+func (g mockGlacierAPI) GetVaultNotificationsRequest(d *glacier.GetVaultNotificationsInput) (*request.Request, *glacier.GetVaultNotificationsOutput) {
 	return g.mockGetVaultNotificationsRequest(d)
 }
 
-func (g glacierAPIMock) GetVaultNotifications(d *glacier.GetVaultNotificationsInput) (*glacier.GetVaultNotificationsOutput, error) {
+func (g mockGlacierAPI) GetVaultNotifications(d *glacier.GetVaultNotificationsInput) (*glacier.GetVaultNotificationsOutput, error) {
 	return g.mockGetVaultNotifications(d)
 }
 
-func (g glacierAPIMock) InitiateJobRequest(i *glacier.InitiateJobInput) (*request.Request, *glacier.InitiateJobOutput) {
+func (g mockGlacierAPI) InitiateJobRequest(i *glacier.InitiateJobInput) (*request.Request, *glacier.InitiateJobOutput) {
 	return g.mockInitiateJobRequest(i)
 }
 
-func (g glacierAPIMock) InitiateJob(i *glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
+func (g mockGlacierAPI) InitiateJob(i *glacier.InitiateJobInput) (*glacier.InitiateJobOutput, error) {
 	return g.mockInitiateJob(i)
 }
 
-func (g glacierAPIMock) InitiateMultipartUploadRequest(i *glacier.InitiateMultipartUploadInput) (*request.Request, *glacier.InitiateMultipartUploadOutput) {
+func (g mockGlacierAPI) InitiateMultipartUploadRequest(i *glacier.InitiateMultipartUploadInput) (*request.Request, *glacier.InitiateMultipartUploadOutput) {
 	return g.mockInitiateMultipartUploadRequest(i)
 }
 
-func (g glacierAPIMock) InitiateMultipartUpload(i *glacier.InitiateMultipartUploadInput) (*glacier.InitiateMultipartUploadOutput, error) {
+func (g mockGlacierAPI) InitiateMultipartUpload(i *glacier.InitiateMultipartUploadInput) (*glacier.InitiateMultipartUploadOutput, error) {
 	return g.mockInitiateMultipartUpload(i)
 }
 
-func (g glacierAPIMock) InitiateVaultLockRequest(i *glacier.InitiateVaultLockInput) (*request.Request, *glacier.InitiateVaultLockOutput) {
+func (g mockGlacierAPI) InitiateVaultLockRequest(i *glacier.InitiateVaultLockInput) (*request.Request, *glacier.InitiateVaultLockOutput) {
 	return g.mockInitiateVaultLockRequest(i)
 }
 
-func (g glacierAPIMock) InitiateVaultLock(i *glacier.InitiateVaultLockInput) (*glacier.InitiateVaultLockOutput, error) {
+func (g mockGlacierAPI) InitiateVaultLock(i *glacier.InitiateVaultLockInput) (*glacier.InitiateVaultLockOutput, error) {
 	return g.mockInitiateVaultLock(i)
 }
 
-func (g glacierAPIMock) ListJobsRequest(l *glacier.ListJobsInput) (*request.Request, *glacier.ListJobsOutput) {
+func (g mockGlacierAPI) ListJobsRequest(l *glacier.ListJobsInput) (*request.Request, *glacier.ListJobsOutput) {
 	return g.mockListJobsRequest(l)
 }
 
-func (g glacierAPIMock) ListJobs(l *glacier.ListJobsInput) (*glacier.ListJobsOutput, error) {
+func (g mockGlacierAPI) ListJobs(l *glacier.ListJobsInput) (*glacier.ListJobsOutput, error) {
 	return g.mockListJobs(l)
 }
 
-func (g glacierAPIMock) ListJobsPages(l *glacier.ListJobsInput, f func(*glacier.ListJobsOutput, bool) bool) error {
+func (g mockGlacierAPI) ListJobsPages(l *glacier.ListJobsInput, f func(*glacier.ListJobsOutput, bool) bool) error {
 	return g.mockListJobsPages(l, f)
 }
 
-func (g glacierAPIMock) ListMultipartUploadsRequest(l *glacier.ListMultipartUploadsInput) (*request.Request, *glacier.ListMultipartUploadsOutput) {
+func (g mockGlacierAPI) ListMultipartUploadsRequest(l *glacier.ListMultipartUploadsInput) (*request.Request, *glacier.ListMultipartUploadsOutput) {
 	return g.mockListMultipartUploadsRequest(l)
 }
 
-func (g glacierAPIMock) ListMultipartUploads(l *glacier.ListMultipartUploadsInput) (*glacier.ListMultipartUploadsOutput, error) {
+func (g mockGlacierAPI) ListMultipartUploads(l *glacier.ListMultipartUploadsInput) (*glacier.ListMultipartUploadsOutput, error) {
 	return g.mockListMultipartUploads(l)
 }
 
-func (g glacierAPIMock) ListMultipartUploadsPages(l *glacier.ListMultipartUploadsInput, f func(*glacier.ListMultipartUploadsOutput, bool) bool) error {
+func (g mockGlacierAPI) ListMultipartUploadsPages(l *glacier.ListMultipartUploadsInput, f func(*glacier.ListMultipartUploadsOutput, bool) bool) error {
 	return g.mockListMultipartUploadsPages(l, f)
 }
 
-func (g glacierAPIMock) ListPartsRequest(l *glacier.ListPartsInput) (*request.Request, *glacier.ListPartsOutput) {
+func (g mockGlacierAPI) ListPartsRequest(l *glacier.ListPartsInput) (*request.Request, *glacier.ListPartsOutput) {
 	return g.mockListPartsRequest(l)
 }
 
-func (g glacierAPIMock) ListParts(l *glacier.ListPartsInput) (*glacier.ListPartsOutput, error) {
+func (g mockGlacierAPI) ListParts(l *glacier.ListPartsInput) (*glacier.ListPartsOutput, error) {
 	return g.mockListParts(l)
 }
 
-func (g glacierAPIMock) ListPartsPages(l *glacier.ListPartsInput, f func(*glacier.ListPartsOutput, bool) bool) error {
+func (g mockGlacierAPI) ListPartsPages(l *glacier.ListPartsInput, f func(*glacier.ListPartsOutput, bool) bool) error {
 	return g.mockListPartsPages(l, f)
 }
 
-func (g glacierAPIMock) ListTagsForVaultRequest(l *glacier.ListTagsForVaultInput) (*request.Request, *glacier.ListTagsForVaultOutput) {
+func (g mockGlacierAPI) ListTagsForVaultRequest(l *glacier.ListTagsForVaultInput) (*request.Request, *glacier.ListTagsForVaultOutput) {
 	return g.mockListTagsForVaultRequest(l)
 }
 
-func (g glacierAPIMock) ListTagsForVault(l *glacier.ListTagsForVaultInput) (*glacier.ListTagsForVaultOutput, error) {
+func (g mockGlacierAPI) ListTagsForVault(l *glacier.ListTagsForVaultInput) (*glacier.ListTagsForVaultOutput, error) {
 	return g.mockListTagsForVault(l)
 }
 
-func (g glacierAPIMock) ListVaultsRequest(l *glacier.ListVaultsInput) (*request.Request, *glacier.ListVaultsOutput) {
+func (g mockGlacierAPI) ListVaultsRequest(l *glacier.ListVaultsInput) (*request.Request, *glacier.ListVaultsOutput) {
 	return g.mockListVaultsRequest(l)
 }
 
-func (g glacierAPIMock) ListVaults(l *glacier.ListVaultsInput) (*glacier.ListVaultsOutput, error) {
+func (g mockGlacierAPI) ListVaults(l *glacier.ListVaultsInput) (*glacier.ListVaultsOutput, error) {
 	return g.mockListVaults(l)
 }
 
-func (g glacierAPIMock) ListVaultsPages(l *glacier.ListVaultsInput, f func(*glacier.ListVaultsOutput, bool) bool) error {
+func (g mockGlacierAPI) ListVaultsPages(l *glacier.ListVaultsInput, f func(*glacier.ListVaultsOutput, bool) bool) error {
 	return g.mockListVaultsPages(l, f)
 }
 
-func (g glacierAPIMock) RemoveTagsFromVaultRequest(r *glacier.RemoveTagsFromVaultInput) (*request.Request, *glacier.RemoveTagsFromVaultOutput) {
+func (g mockGlacierAPI) RemoveTagsFromVaultRequest(r *glacier.RemoveTagsFromVaultInput) (*request.Request, *glacier.RemoveTagsFromVaultOutput) {
 	return g.mockRemoveTagsFromVaultRequest(r)
 }
 
-func (g glacierAPIMock) RemoveTagsFromVault(r *glacier.RemoveTagsFromVaultInput) (*glacier.RemoveTagsFromVaultOutput, error) {
+func (g mockGlacierAPI) RemoveTagsFromVault(r *glacier.RemoveTagsFromVaultInput) (*glacier.RemoveTagsFromVaultOutput, error) {
 	return g.mockRemoveTagsFromVault(r)
 }
 
-func (g glacierAPIMock) SetDataRetrievalPolicyRequest(s *glacier.SetDataRetrievalPolicyInput) (*request.Request, *glacier.SetDataRetrievalPolicyOutput) {
+func (g mockGlacierAPI) SetDataRetrievalPolicyRequest(s *glacier.SetDataRetrievalPolicyInput) (*request.Request, *glacier.SetDataRetrievalPolicyOutput) {
 	return g.mockSetDataRetrievalPolicyRequest(s)
 }
 
-func (g glacierAPIMock) SetDataRetrievalPolicy(s *glacier.SetDataRetrievalPolicyInput) (*glacier.SetDataRetrievalPolicyOutput, error) {
+func (g mockGlacierAPI) SetDataRetrievalPolicy(s *glacier.SetDataRetrievalPolicyInput) (*glacier.SetDataRetrievalPolicyOutput, error) {
 	return g.mockSetDataRetrievalPolicy(s)
 }
 
-func (g glacierAPIMock) SetVaultAccessPolicyRequest(s *glacier.SetVaultAccessPolicyInput) (*request.Request, *glacier.SetVaultAccessPolicyOutput) {
+func (g mockGlacierAPI) SetVaultAccessPolicyRequest(s *glacier.SetVaultAccessPolicyInput) (*request.Request, *glacier.SetVaultAccessPolicyOutput) {
 	return g.mockSetVaultAccessPolicyRequest(s)
 }
 
-func (g glacierAPIMock) SetVaultAccessPolicy(s *glacier.SetVaultAccessPolicyInput) (*glacier.SetVaultAccessPolicyOutput, error) {
+func (g mockGlacierAPI) SetVaultAccessPolicy(s *glacier.SetVaultAccessPolicyInput) (*glacier.SetVaultAccessPolicyOutput, error) {
 	return g.mockSetVaultAccessPolicy(s)
 }
 
-func (g glacierAPIMock) SetVaultNotificationsRequest(s *glacier.SetVaultNotificationsInput) (*request.Request, *glacier.SetVaultNotificationsOutput) {
+func (g mockGlacierAPI) SetVaultNotificationsRequest(s *glacier.SetVaultNotificationsInput) (*request.Request, *glacier.SetVaultNotificationsOutput) {
 	return g.mockSetVaultNotificationsRequest(s)
 }
 
-func (g glacierAPIMock) SetVaultNotifications(s *glacier.SetVaultNotificationsInput) (*glacier.SetVaultNotificationsOutput, error) {
+func (g mockGlacierAPI) SetVaultNotifications(s *glacier.SetVaultNotificationsInput) (*glacier.SetVaultNotificationsOutput, error) {
 	return g.mockSetVaultNotifications(s)
 }
 
-func (g glacierAPIMock) UploadArchiveRequest(u *glacier.UploadArchiveInput) (*request.Request, *glacier.ArchiveCreationOutput) {
+func (g mockGlacierAPI) UploadArchiveRequest(u *glacier.UploadArchiveInput) (*request.Request, *glacier.ArchiveCreationOutput) {
 	return g.mockUploadArchiveRequest(u)
 }
 
-func (g glacierAPIMock) UploadArchive(u *glacier.UploadArchiveInput) (*glacier.ArchiveCreationOutput, error) {
+func (g mockGlacierAPI) UploadArchive(u *glacier.UploadArchiveInput) (*glacier.ArchiveCreationOutput, error) {
 	return g.mockUploadArchive(u)
 }
 
-func (g glacierAPIMock) UploadMultipartPartRequest(u *glacier.UploadMultipartPartInput) (*request.Request, *glacier.UploadMultipartPartOutput) {
+func (g mockGlacierAPI) UploadMultipartPartRequest(u *glacier.UploadMultipartPartInput) (*request.Request, *glacier.UploadMultipartPartOutput) {
 	return g.mockUploadMultipartPartRequest(u)
 }
 
-func (g glacierAPIMock) UploadMultipartPart(u *glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
+func (g mockGlacierAPI) UploadMultipartPart(u *glacier.UploadMultipartPartInput) (*glacier.UploadMultipartPartOutput, error) {
 	return g.mockUploadMultipartPart(u)
 }
 
-func (g glacierAPIMock) WaitUntilVaultExists(d *glacier.DescribeVaultInput) error {
+func (g mockGlacierAPI) WaitUntilVaultExists(d *glacier.DescribeVaultInput) error {
 	return g.mockWaitUntilVaultExists(d)
 }
 
-func (g glacierAPIMock) WaitUntilVaultNotExists(d *glacier.DescribeVaultInput) error {
+func (g mockGlacierAPI) WaitUntilVaultNotExists(d *glacier.DescribeVaultInput) error {
 	return g.mockWaitUntilVaultNotExists(d)
 }