+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Encrypt/decrypt backup with a shared secret
 - Encryption data authentication (HMAC-SHA256)
+- Send report with the scheduler actions periodically
 
 ### Fixed
 - Add sample configuration file to deb and txz packages